backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.0.69
langchain==0.1.0
langchain-openai==0.0.2rc1
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
